+++ lib/src/fragments.dart
@@ -1,3 +1,5 @@
+import 'dart:collection';
+
 import 'package:flutter/widgets.dart';
 
 import 'utils.dart';
@@ -8,52 +10,35 @@ mixin Fragments<W extends StatefulWidget> on State<W> {
   /// [builder] will be called only when [deps] is different (not shallowly equal)
   /// from the previous pass.
   T fragment<T>(T builder(T prev), {Iterable deps = const [], Key key}) {
-    // disable nested fragment() calls for now
-    assert(_fragmentDepth.current == 0);
-    final isAnonymous = key == null;
+    if (root.isLocked.now) {
+      final parent = root.path.last;
+      final self = parent.children[parent.childCursor.now];
+      assert(self.childCursor.now == 0);
 
-    final bool hasCache = isAnonymous
-        ? _subtreeCursor.current < _anonymousSubtrees.length &&
-            _anonymousSubtrees[_subtreeCursor.current].value is T
-        : _namedSubtrees.containsKey(key) && _namedSubtrees[key].value is T;
-
-    final T cachedValue = hasCache
-        ? isAnonymous
-            ? _anonymousSubtrees[_subtreeCursor.current].value as T
-            : _namedSubtrees[key].value as T
-        : null;
-
-    // try use existing cache
-    if (hasCache) {
-      final cachedDeps = isAnonymous
-          ? _anonymousSubtrees[_subtreeCursor.current].deps
-          : _namedSubtrees[key].deps;
-
-      if (shallowEquals(cachedDeps, deps)) {
-        if (isAnonymous) _subtreeCursor.current++;
-        return cachedValue;
-      }
-    }
-
-    // update cache
-    try {
-      _fragmentDepth.current++;
-      // rebuild subtree
-      final newWidget = builder(cachedValue);
-
-      // save subtree to cache
-      if (isAnonymous) {
-        if (_subtreeCursor.current >= _anonymousSubtrees.length)
-          _anonymousSubtrees.length = _subtreeCursor.current + 1;
-        _anonymousSubtrees[_subtreeCursor.current] = _Subtree(newWidget, deps);
-        _subtreeCursor.current++;
-      } else {
-        _namedSubtrees[key] = _Subtree(newWidget, deps);
+      if (self.item.value is T && !shallowEquals(self.item.deps, deps)) {
+        root.path.addLast(self);
+        self.item = _CacheItem(builder(self.item.value), deps);
+        assert(identical(root.path.last, self));
+        root.path.removeLast();
       }
 
-      return newWidget;
-    } finally {
-      _fragmentDepth.current--;
+      self.childCursor.now = 0;
+      parent.childCursor.now++;
+      return self.item.value;
+    } else {
+      final parent = root.path.last;
+      final self = _CacheNode();
+      parent.children.add(self);
+      assert(identical(parent.children[parent.childCursor.now], self));
+
+      root.path.addLast(self);
+      self.item = _CacheItem(builder(null), deps);
+      assert(identical(root.path.last, self));
+      root.path.removeLast();
+
+      self.childCursor.now = 0;
+      parent.childCursor.now++;
+      return self.item.value;
     }
   }
 
@@ -61,47 +46,81 @@ mixin Fragments<W extends StatefulWidget> on State<W> {
   @mustCallSuper
   void didUpdateWidget(oldWidget) {
     super.didUpdateWidget(oldWidget);
-    _subtreeCursor.current = 0;
+    root.childCursor.now = 0;
+    root.isLocked.now = true;
   }
 
   @override
   @mustCallSuper
   void didChangeDependencies() {
     super.didChangeDependencies();
-    _subtreeCursor.current = 0;
+    if (didInit.now) {
+      root.childCursor.now = 0;
+      root.isLocked.now = true;
+    } else {
+      didInit.now = true;
+    }
   }
 
   @override
   @mustCallSuper
   void setState(VoidCallback fn) {
     super.setState(fn);
-    _subtreeCursor.current = 0;
+    root.childCursor.now = 0;
+    root.isLocked.now = true;
   }
 
   @override
   void reassemble() {
-    _subtreeCursor.current = 0;
-    _fragmentDepth.current = 0;
-    _namedSubtrees.clear();
-    _anonymousSubtrees.clear();
     super.reassemble();
+    root.reset();
   }
 
-  final _Ref<int> _subtreeCursor = _Ref(0);
-  final Map<Key, _Subtree> _namedSubtrees = <Key, _Subtree>{};
-  final List<_Subtree> _anonymousSubtrees = <_Subtree>[];
-  final _Ref<int> _fragmentDepth = _Ref(0);
+  final root = _CacheRoot();
+  final didInit = _Ref(false);
 }
 
-@immutable
-class _Subtree {
+class _CacheItem {
   final Object value;
   final Iterable deps;
 
-  _Subtree(this.value, this.deps);
+  _CacheItem(this.value, this.deps);
+}
+
+class _CacheRoot with _HasChildren {
+  final Queue<_HasChildren> path = Queue();
+
+  final isLocked = _Ref(false);
+
+  void reset() {
+    super.reset();
+    isLocked.now = false;
+    path.clear();
+    path.addLast(this);
+  }
+
+  _CacheRoot() {
+    path.addLast(this);
+  }
+}
+
+class _CacheNode with _HasChildren {
+  _CacheItem item;
+}
+
+mixin _HasChildren {
+  @protected
+  final children = List<_CacheNode>();
+  final childCursor = _Ref(0);
+
+  @mustCallSuper
+  void reset() {
+    children.clear();
+    childCursor.now = 0;
+  }
 }
 
 class _Ref<T> {
-  _Ref(T init) : current = init;
-  T current;
+  _Ref(T init) : now = init;
+  T now;
 }