+++ CHANGELOG.md
@@ -35,3 +35,6 @@
 ## [0.3.0]
 
 * Provide previous value to builder
+* Make nested `fragment()` calls act like React's `useMemo` 
+* Fix nested fragments support
+* Remove named fragments support